remove encrypt at rest code
