+++ docker-compose.yml
@@ -27,9 +27,6 @@ services:
           INGEST_PASS: "${INGEST_PASS}"
           LOG_LEVEL: "${LOG_LEVEL}"
           SECRET_KEY: "${SECRET_KEY}"
-          ENCRYPT_AT_REST: "${ENCRYPT_AT_REST}"
-          ENCRYPT_KEY: "${ENCRYPT_KEY}"
-
         ports:
           - "5000:5000"
     
@@ -46,8 +43,6 @@ services:
           VALIDATOR_USER: "${VALIDATOR_USER}"
           VALIDATOR_PASS: "${VALIDATOR_PASS}"
           LOG_LEVEL: "${LOG_LEVEL}"
-          ENCRYPT_AT_REST: "${ENCRYPT_AT_REST}"
-          ENCRYPT_KEY: "${ENCRYPT_KEY}"
 
     writer:
         build: 
@@ -66,17 +61,4 @@ services:
           DB_NAME: "${DB_NAME}"
           DB_USERNAME: "${DB_USERNAME}"
           DB_PASSWORD: "${DB_PASSWORD}"
-          ENCRYPT_AT_REST: "${ENCRYPT_AT_REST}"
-          ENCRYPT_KEY: "${ENCRYPT_KEY}"
-
-    paybc_api:
-        build:
-          context: python
-          dockerfile: paybc_api/Dockerfile-local
-        environment:
-          PAYBC_FLASK_SECRET: "${PAYBC_FLASK_SECRET}"
-          PAYBC_CLIENT_ID: "${PAYBC_CLIENT_ID}"
-          PAYBC_CLIENT_SECRET: "${PAYBC_CLIENT_SECRET}"
-
-        ports:
-          - "5001:5001"
+
