+++ python/common/config.py
@@ -10,8 +10,6 @@ class Config:
     RETRY_DELAY                 = os.getenv('RETRY_DELAY', 30)
     LOG_LEVEL                   = os.environ.get('LOG_LEVEL', 'INFO').upper()
     RABBITMQ_MESSAGE_ENCODE     = os.getenv('RABBITMQ_MESSAGE_ENCODE', 'utf-8')
-    ENCRYPT_AT_REST             = (os.getenv('ENCRYPT_AT_REST', 'False').upper() == "TRUE")
-    ENCRYPT_KEY                 = os.getenv('ENCRYPT_KEY')
