+++ python/ingestor/ingest.py
@@ -1,7 +1,6 @@
 from python.ingestor.config import Config
 from python.common.rabbitmq import RabbitMQ
-from python.common.message_factory import MessageFactory
-from flask import request, jsonify, Response
+from flask import request, jsonify, Response, json
 from flask_api import FlaskAPI
 import logging
 
@@ -18,16 +17,10 @@ rabbit_mq = RabbitMQ(
     Config.MAX_CONNECTION_RETRIES,
     Config.RETRY_DELAY)
 
-message = MessageFactory.get_message(
-    Config.ENCRYPT_AT_REST,
-    Config.RABBITMQ_MESSAGE_ENCODE,
-    Config.LOG_LEVEL,
-    Config.ENCRYPT_KEY)
-
 
 @application.route('/v1/publish/event', methods=["POST"])
 def create():
-    if rabbit_mq.publish(Config.WRITE_QUEUE, message.encode_ingested_message(request.json)):
+    if rabbit_mq.publish(Config.WRITE_QUEUE, json.dumps(request.json)):
         return jsonify(request.json), 200
     else:
         return Response('Unavailable', 500, mimetype='application/json')
