+++ python/ingestor/requirements.txt
@@ -1,5 +1,4 @@
 Click==7.0
-cryptography==2.9.2
 Flask==1.1.1
 Flask-API==2.0
 gunicorn==20.0.4
