+++ python/validator/listener.py
@@ -1,8 +1,9 @@
 from python.validator.config import Config
 from python.validator.validator import Validate
 from python.common.rabbitmq import RabbitMQ
-from python.common.message_factory import MessageFactory
+from python.common.helper import Helper
 import logging
+import json
 
 
 class Listener:
@@ -13,12 +14,11 @@ class Listener:
          - determines whether the message is valid or not valid,
          - writes the message to a valid or not valid queue
     """
-    def __init__(self, config, validator, rabbit_writer, rabbit_listener, message):
+    def __init__(self, config, validator, rabbit_writer, rabbit_listener):
         self.validator = validator
         self.config = config
         self.writer = rabbit_writer
         self.listener = rabbit_listener
-        self.message = message
 
         logging.basicConfig(level=config.LOG_LEVEL)
         logging.warning('*** validator initialized  ***')
@@ -34,20 +34,21 @@ class Listener:
     def callback(self, ch, method, properties, body):
         logging.info('message received; callback invoked')
 
-        message_dict = self.message.decode_ingested_message(body)
+        # convert body (in bytes) to string
+        message = body.decode(self.config.RABBITMQ_MESSAGE_ENCODE)
+        message_dict = json.loads(message)
 
-        result = self.validator.validate(message_dict)
-        if result['isSuccess']:
-            # Validation SUCCESSFUL
+        validation_result = self.validator.validate(message_dict)
+        if validation_result['isSuccess']:
+            # Validation was SUCCESSFUL
             logging.info("write to: " + self.config.VALID_QUEUE)
-            if self.writer.publish(self.config.VALID_QUEUE, self.message.encode_validated_message(message_dict)):
+            if self.writer.publish(self.config.VALID_QUEUE, message):
                 ch.basic_ack(delivery_tag=method.delivery_tag)
         else:
             # Validation FAILED
+            message_with_errors_appended = Helper.add_error_to_message(message_dict, validation_result['description'])
             logging.info("write to: " + self.config.FAIL_QUEUE)
-            if self.writer.publish(
-                    self.config.FAIL_QUEUE,
-                    self.message.encode_validated_message(message_dict, result)):
+            if self.writer.publish(self.config.FAIL_QUEUE, json.dumps(message_with_errors_appended)):
                 ch.basic_ack(delivery_tag=method.delivery_tag)
 
 
@@ -68,10 +69,5 @@ if __name__ == "__main__":
             Config.RABBITMQ_URL,
             Config.LOG_LEVEL,
             Config.MAX_CONNECTION_RETRIES,
-            Config.RETRY_DELAY),
-        MessageFactory.get_message(
-            Config.ENCRYPT_AT_REST,
-            Config.RABBITMQ_MESSAGE_ENCODE,
-            Config.LOG_LEVEL,
-            Config.ENCRYPT_KEY)
+            Config.RETRY_DELAY)
     ).main()
