+++ python/validator/requirements.txt
@@ -1,3 +1,2 @@
 Cerberus==1.3.2
-cryptography==2.9.2
 pika==1.1.0
