+++ python/writer/requirements.txt
@@ -1,4 +1,3 @@
-cryptography==2.9.2
 pika==1.1.0
 pyodbc==4.0.28
 sql==0.4.0
